+++ packages/stream_chat_persistence/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Upcoming
+
+- Reintroduce support for experimental indexedDB on Web.
+
 ## 5.0.0
 
 - Included the changes from version [4.3.0](#430) and [4.4.0](#440).